style: provide folder structure
