+++ packages/stream_chat_v1/lib/pages/channel_media_display_screen.dart
@@ -1,4 +1,4 @@
-import 'package:example/localizations.dart';
+import 'package:example/utils/localizations.dart';
 import 'package:example/routes/routes.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
