+++ packages/stream_chat_v1/lib/pages/chat_info_screen.dart
@@ -1,11 +1,10 @@
-import 'package:example/localizations.dart';
+import 'package:example/pages/channel_file_display_screen.dart';
+import 'package:example/utils/localizations.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-
-import 'channel_file_display_screen.dart';
-import 'channel_media_display_screen.dart';
-import 'pinned_messages_screen.dart';
+import 'package:example/pages/channel_media_display_screen.dart';
+import 'package:example/pages/pinned_messages_screen.dart';
 
 /// Detail screen for a 1:1 chat correspondence
 class ChatInfoScreen extends StatefulWidget {
