+++ packages/stream_chat_v1/lib/pages/home_page.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import 'package:example/channel_page.dart';
-import 'package:example/notifications_service.dart';
+import 'package:example/pages/channel_page.dart';
+import 'package:example/utils/notifications_service.dart';
 import 'package:example/routes/app_routes.dart';
 import 'package:example/routes/routes.dart';
 import 'package:flutter/material.dart';
