+++ packages/stream_chat_v1/lib/pages/splash_screen.dart
