+++ packages/stream_chat_v1/lib/pages/thread_page.dart
