+++ packages/stream_chat_v1/lib/routes/routes.dart
@@ -1,5 +1,5 @@
-/// Define all the route names here
-class Routes {
+/// Application routes
+abstract class Routes {
   static const String APP = '/app';
   static const String HOME = '/home';
   static const String CHOOSE_USER = '/choose_user';
