+++ packages/stream_chat_v1/lib/widgets/search_text_field.dart
